requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
whitenoise==6.11.0
